requirements.txt
langchain
python-dotenv
PyPDF2
flask 
pinecone
pinecone-client
langchain_google_genai

-e .